fix: handle undefined paths in getModifiedTime
When building projects in `tsc` build mode (with `typescript.build` set
to `true`), some paths passed to `getModifiedTime` will be `undefined`.
Previously, this would cause an error and prevent type-checking (see
https://github.com/TypeStrong/fork-ts-checker-webpack-plugin/issues/855).
With these changes, the modified time will just be returned as
`undefined` in such cases, which allows type checking to complete
successfully.